// medical-clinics-frontend/src/app/services/models/user-account-edit-request.ts
/* tslint:disable */
/* eslint-disable */

/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface UserAccountEditRequest {
  address: string | null;
  city: string | null;
  country: string | null;
  email: string;
  firstName: string;
  id: string;
  lastName: string;
  newPassword: string | null;
  oldPassword: string | null;
  phone: string | null;
}